Rename testutil into pkitest for consistency with other packages
No functional changes
