+++ pkg/creator/ssam/BUILD.bazel
@@ -36,7 +36,7 @@ go_test(
     deps = [
         "//pkg/util/httptest:go_default_library",
         "//pkg/util/httputil:go_default_library",
-        "//pkg/util/pkiutil/testutil:go_default_library",
+        "//pkg/util/pkiutil/pkitest:go_default_library",
         "//pkg/util/testutil:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
         "//vendor/github.com/stretchr/testify/require:go_default_library",
