+++ pkg/servicecentral/BUILD.bazel
@@ -47,7 +47,7 @@ go_test(
         "//pkg/util/httptest:go_default_library",
         "//pkg/util/httputil:go_default_library",
         "//pkg/util/pkiutil:go_default_library",
-        "//pkg/util/pkiutil/testutil:go_default_library",
+        "//pkg/util/pkiutil/pkitest:go_default_library",
         "//pkg/util/testutil:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
